requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==18.1.0
beautifulsoup4==4.6.1
certifi==2018.1.18
chardet==3.0.4
Click==7.0
colorama==0.3.9
configparser==3.5.0
crayons==0.1.2
cx-Oracle==6.3.1
cycler==0.10.0
decorator==4.4.0
docopt==0.6.2
entrypoints==0.3
enum34==1.1.6
flake8==3.5.0
flake8-builtins==1.4.1
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
Flask==1.0.2
idna==2.6
importlib-metadata==0.18
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.23
jsonpath-rw==1.4.0
jsonpointer==2.0
jsonschema==3.0.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
MonthDelta==0.9.1
more-itertools==6.0.0
mysql-connector-python==8.0.15
namedlist==1.7
numpy==1.16.2
opencv-python==4.0.1.23
ordereddict==1.1
packaging==19.0
Pillow==5.2.0
pipreqs==0.4.9
pluggy==0.6.0
ply==3.11
psycopg2==2.8.2
py==1.8.0
pycodestyle==2.3.1
pydocstyle==3.0.0
pyflakes==1.6.0
PyHamcrest==1.9.0
pylint==2.3.1
pyodbc==4.0.26
pyparsing==2.4.0
pypng==0.0.19
PyQRCode==1.2.1
pyrsistent==0.14.11
pytest==3.6.1
pytest-allure-adaptor==1.7.10
pytest-testrail==1.1.6
python-dateutil==2.8.0
pyzbar==0.1.8
qrtools==0.0.2
redis==3.2.1
requests==2.18.4
robotframework==3.0.2
robotframework-databaselibrary==1.2
robotframework-httplibrary==0.4.2
robotframework-pabot==0.44
robotframework-requests==0.5.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1
robotremoteserver==1.1
sauceclient==1.0.0
selenium==3.14.0
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
soupsieve==1.9
terminaltables==3.1.0
typed-ast==1.4.0
urllib3==1.22
visual-logging==1.0
vk==2.0.2
vk-api==11.4.0
waitress==1.1.0
wcwidth==0.1.7
webdriver-manager==1.7
WebOb==1.8.2
WebTest==2.0.30
Werkzeug==0.15.2
wrapt==1.11.2
yarg==0.1.9
zipp==0.5.1
